refactor: move to a class-based arch
